Updated colors and fixed global delete

+++ kanban.client/src/components/BoardComponent.vue
@@ -1,14 +1,15 @@
 <template>
-  <div class="board-component col-3 position-relative">
+  <div class="board-component col-3 p-0 position-relative" v-if="state.account">
+    <img class="img-icon icon-overlay my-auto" :src="state.account.picture" :title="(state.account.name).split('@')[0]">
     <button type="button" class="btn btn-sm btn-outline-danger btn-size btn-overlay p-0">
-      <i class="fas fa-times" @click="deleteBoard(boardProp.id)"></i>
+      <i class="fas fa-times" @click="deleteBoard(boardProp.id, boardProp.creatorId)"></i>
     </button>
     <router-link :to="{ name: 'Board', params: { id:boardProp.id } }">
       <!-- ROUTER LINK WRAPPING BOARD -->
       <div class="row bg-light shadow m-3">
-        <div class="col-12 text-center hover-light py-2">
+        <div class="col-12 text-center hover-light">
           <h3><u>{{ boardProp.title }}</u></h3>
-          <img class="img-fluid mt-1 mb-2" :src="boardProp.imgUrl" alt="">
+          <img class="img-fluid mt-1 mb-2 pb-2" :src="boardProp.imgUrl" alt="">
         </div>
       </div>
       <!-- ROUTER LINK WRAPPING BOARD -->
@@ -18,7 +19,9 @@
 
 <script>
 import { boardsService } from '../services/BoardsService'
+import { AppState } from '../AppState'
 import Notification from '../utils/Notification'
+import { computed, reactive } from 'vue'
 
 export default {
   name: 'BoardComponent',
@@ -29,12 +32,25 @@ export default {
     }
   },
   setup() {
+    const state = reactive({
+      account: computed(() => AppState.account)
+    })
     return {
-      async deleteBoard(id) {
-        try {
-          await boardsService.deleteBoard(id)
-        } catch (error) {
-          Notification.toast('Error: ' + error, 'error')
+      state,
+      async deleteBoard(id, creatorId) {
+        if (state.account.id !== creatorId) {
+          Notification.toast("Denied! That's not yours!", 'danger')
+        } else {
+          if (await Notification.confirmAction()) {
+            try {
+              await boardsService.deleteBoard(id)
+              Notification.toast('Deleted! ', 'error')
+            } catch (error) {
+              Notification.toast('Error: ' + error, 'error')
+            }
+          } else {
+            Notification.toast('No worries!', 'success')
+          }
         }
       }
     }
